pyproject.toml
[tool.poetry]
name = "pyrgbd"
version = "0.1.0"
description = ""
authors = ["Hanseul Jun <hanseuljun@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
cffi = "^1.15.1"
numpy = "^1.23.2"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
